pyproject.toml
[project]
name = "salo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "google-cloud-firestore>=2.22.0",
    "gunicorn>=23.0.0",
]
